upgrade spring boot version to 3.4.4 and fix all dependencies due to the upgrading, and change the autumn version to 3.0.0

+++ autumn-lib/src/main/java/cn/org/autumn/utils/PageUtils.java
@@ -1,6 +1,6 @@
 package cn.org.autumn.utils;
 
-import com.baomidou.mybatisplus.plugins.Page;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 
 import java.io.Serializable;
 import java.util.List;
@@ -41,10 +41,10 @@ public class PageUtils implements Serializable {
 	 */
 	public PageUtils(Page<?> page) {
 		this.list = page.getRecords();
-		this.totalCount = page.getTotal();
-		this.pageSize = page.getSize();
-		this.currPage = page.getCurrent();
-		this.totalPage = page.getPages();
+		this.totalCount = (int) page.getTotal();
+		this.pageSize = (int) page.getSize();
+		this.currPage = (int) page.getCurrent();
+		this.totalPage = (int) page.getPages();
 	}
 
 	public int getTotalCount() {
